build: enforce clang-based warning checks on every C compile
cproc explicitly ignores -W flags (driver.c:549 "ignore warning flag"),
so cc65816 has been silently swallowing every diagnostic the SDK ever
should have produced. Run clang -fsyntax-only -Wall -Wextra -Werror as
a sibling step in the .c.o rule of make/common.mk; SKIP_LINT=1 is the
escape hatch for environments without clang.

The check found 12 latent warnings — all fixed in this commit:

Real bugs:
  - lib/source/hdma.c:89 — sine_quarter[63] = 256 truncates to 0 in u8,
    so HDMA sine effects had a 1-pixel notch at sin(90°). Cap the table
    at 255 (1/256 precision loss is imperceptible).
  - games/tetris/render.c:346 — `u8 tile` couldn't hold TILE_BORDER_H |
    PAL1 = 0x409, dropping the palette bits silently and rendering the
    line-clear flash with the wrong palette. Promote to u16.

Cleanup (unused variables / functions, no behaviour change):
  - 4 unused locals stripped (collision_demo, dynamic_sprite, tetris/board)
  - 2 trivial dead wrappers removed (tetris/hud writestring_bg2,
    dynamic_map refresh)
  - convertC64Sprite kept with __attribute__((unused)) — referenced from
    a doc comment as the canonical reference for the rockford_8bpp data

The lint flags disable -Wpointer-to-int-cast / -Wint-to-pointer-cast
(SNES has 16-bit pointers, not 64-bit; clang's host model gets the
casts wrong) and -Wunused-parameter (object-engine callbacks have a
fixed ABI even when individual handlers ignore some args).

Closes P2.3 of the remediation plan.

+++ examples/games/tetris/render.c
@@ -343,7 +343,7 @@ void renderClearNextPiece(void) {
 
 void renderLineClearFlash(LineClearResult *result, u8 frame) {
     u8 i, c, r;
-    u8 tile;
+    u16 tile;
 
     tile = (frame & 1) ? (TILE_BORDER_H | PAL1) : TILE_EMPTY;
